plant-disease-detection/requirements.txt
tensorflow
fastapi
uvicorn
python-multipart
Pillow
numpy